refactor(template-generator): create an unified  to have a consistent data format and unified output
